backend/SRLMS/requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.40.0
django-cors-headers==3.1.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.10.2
drf-writable-nested==0.5.1
idna==2.8
oauthlib==3.1.0
Pillow==6.1.0
psycopg2-binary==2.8.3
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
xlrd==1.2.0
